A SECURE WEB INTERFACE FOR FACE BASED IDENTIFICATION OF MISSING CHILDREN THROUGH LOGGING AND SEARCHING
The count of missing children is growing exponentially with every single passing day. It is implausible that a minimum count of 175 children are being missed every day. Keeping this in mind, the proposed work focuses on an end-to-end web interface for finding missing children through secure logging and searching for missing children using face recognition. Face recognition is being carried out by matching the query face image with the template image found in the database using the VGG-Face algorithm from the Deep Face library; using this, the distance of differences between images is carried out through the Euclidean loss method. This process begins with an admin module, which is the police authority that has the authority to upload the missing person's details with authenticated login credentials. Then, if a volunteer finds a child who was found, then he can click the photo of the child and put it in the web interface where the face is compared with already present faces in the databases uploaded by the police community. If there is a match, then a notification will be received by the police and sent to the family member’s email ID. So that the child’s parents can also know with whom their child is found. A local database is used in the project to store the details of the missing child, such as the guardian's name, address, email ID, and phone number. The total web framework is being carried out by Streamlit for designing the web pages.